[NFC][ESI][MSFT] Move AppIDs over to ESI
Since AppIDs serve as system identifiers, they classify as a system
construction feature. So they should be in the ESI dialect, which will
be using them shortly.
