+++ frontends/PyCDE/src/system.py
@@ -67,7 +67,7 @@ class System:
     self._instance_roots: dict[(Module, str), InstanceHierarchyRoot] = {}
 
     self._placedb: PlacementDB = None
-    self._appid_index: msft.AppIDIndex = None
+    self._appid_index: esi.AppIDIndex = None
 
     # The set of all files generated by PyCDE.
     self.files: Set[os.PathLike] = set()
@@ -226,7 +226,7 @@ class System:
         m.generate()
         i += 1
 
-    self._appid_index = msft.AppIDIndex(self.mod.operation)
+    self._appid_index = esi.AppIDIndex(self.mod.operation)
 
   def get_instance(self,
                    mod_cls: object,
