+++ include/circt/Dialect/ESI/ESI.td
@@ -19,6 +19,7 @@ include "mlir/IR/OpBase.td"
 
 include "circt/Dialect/ESI/ESIDialect.td"
 
+include "circt/Dialect/ESI/ESIAppID.td"
 include "circt/Dialect/ESI/ESIInterfaces.td"
 include "circt/Dialect/ESI/ESITypes.td"
 include "circt/Dialect/ESI/ESIChannels.td"
