+++ include/circt/Dialect/MSFT/MSFTAttributes.td
@@ -73,36 +73,3 @@ def LocationVector : MSFT_Attr<"LocationVector"> {
   let hasCustomAssemblyFormat = 1;
   let genVerifyDecl = 1;
 }
-
-def AppIDAttr : MSFT_Attr<"AppID"> {
-  let summary = "An application relevant instance identifier";
-  let description = [{
-    Identifies an instance which is visible through multiple hierarchy levels.
-    Indended to make locating an instance easier in the instance hierarchy.
-  }];
-
-  let parameters = (ins "StringAttr":$name, "uint64_t":$index);
-  let mnemonic = "appid";
-  let assemblyFormat = [{
-    `<` $name `[` $index `]` `>`
-  }];
-
-  let extraClassDeclaration = [{
-    static constexpr StringRef AppIDAttrName = "msft.appid";
-  }];
-}
-
-def AppIDPathAttr : MSFT_Attr<"AppIDPath"> {
-  let summary = "An application-specific hierarchical path through a design";
-  let description = [{
-    A list of AppID components which specifies a specific dynamic instance
-    in the design.
-  }];
-
-  let parameters = (ins "FlatSymbolRefAttr":$root,
-                        ArrayRefParameter<"AppIDAttr">:$path);
-  let mnemonic = "appid_path";
-  let assemblyFormat = [{
-    `<` $root `[` $path `]` `>`
-  }];
-}
