+++ lib/CAPI/Dialect/MSFT.cpp
@@ -4,7 +4,6 @@
 
 #include "circt-c/Dialect/MSFT.h"
 #include "circt/Dialect/HW/HWSymCache.h"
-#include "circt/Dialect/MSFT/AppID.h"
 #include "circt/Dialect/MSFT/DeviceDB.h"
 #include "circt/Dialect/MSFT/ExportTcl.h"
 #include "circt/Dialect/MSFT/MSFTAttributes.h"
@@ -231,95 +230,4 @@ intptr_t circtMSFTLocationVectorAttrGetNumElements(MlirAttribute attr) {
 MlirAttribute circtMSFTLocationVectorAttrGetElement(MlirAttribute attr,
                                                     intptr_t pos) {
   return wrap(unwrap(attr).cast<LocationVectorAttr>().getLocs()[pos]);
-}
-
-bool circtMSFTAttributeIsAnAppIDAttr(MlirAttribute attr) {
-  return unwrap(attr).isa<AppIDAttr>();
-}
-
-MlirAttribute circtMSFTAppIDAttrGet(MlirContext ctxt, MlirStringRef name,
-                                    uint64_t index) {
-  return wrap(AppIDAttr::get(
-      unwrap(ctxt), StringAttr::get(unwrap(ctxt), unwrap(name)), index));
-}
-MlirStringRef circtMSFTAppIDAttrGetName(MlirAttribute attr) {
-  return wrap(unwrap(attr).cast<AppIDAttr>().getName().getValue());
-}
-uint64_t circtMSFTAppIDAttrGetIndex(MlirAttribute attr) {
-  return unwrap(attr).cast<AppIDAttr>().getIndex();
-}
-
-bool circtMSFTAttributeIsAnAppIDPathAttr(MlirAttribute attr) {
-  return isa<AppIDPathAttr>(unwrap(attr));
-}
-
-MlirAttribute circtMSFTAppIDAttrPathGet(MlirContext ctxt, MlirAttribute root,
-                                        intptr_t numElements,
-                                        MlirAttribute const *cElements) {
-  SmallVector<AppIDAttr, 8> elements;
-  for (intptr_t i = 0; i < numElements; ++i)
-    elements.push_back(cast<AppIDAttr>(unwrap(cElements[i])));
-  return wrap(AppIDPathAttr::get(
-      unwrap(ctxt), cast<FlatSymbolRefAttr>(unwrap(root)), elements));
-}
-MlirAttribute circtMSFTAppIDAttrPathGetRoot(MlirAttribute attr) {
-  return wrap(cast<AppIDPathAttr>(unwrap(attr)).getRoot());
-}
-uint64_t circtMSFTAppIDAttrPathGetNumComponents(MlirAttribute attr) {
-  return cast<AppIDPathAttr>(unwrap(attr)).getPath().size();
-}
-MlirAttribute circtMSFTAppIDAttrPathGetComponent(MlirAttribute attr,
-                                                 uint64_t index) {
-  return wrap(cast<AppIDPathAttr>(unwrap(attr)).getPath()[index]);
-}
-
-//===----------------------------------------------------------------------===//
-// AppID
-//===----------------------------------------------------------------------===//
-
-DEFINE_C_API_PTR_METHODS(CirctMSFTAppIDIndex, circt::msft::AppIDIndex)
-
-/// Create an index of appids through which to do appid lookups efficiently.
-MLIR_CAPI_EXPORTED CirctMSFTAppIDIndex
-circtMSFTAppIDIndexGet(MlirOperation root) {
-  auto *idx = new AppIDIndex(unwrap(root));
-  if (idx->isValid())
-    return wrap(idx);
-  return CirctMSFTAppIDIndex{nullptr};
-}
-
-/// Free an AppIDIndex.
-MLIR_CAPI_EXPORTED void circtMSFTAppIDIndexFree(CirctMSFTAppIDIndex index) {
-  delete unwrap(index);
-}
-
-/// Lookup a DynamicInstanceOp from an appid path.
-MLIR_CAPI_EXPORTED MlirOperation circtMSFTAppIDIndexGetInstance(
-    CirctMSFTAppIDIndex index, MlirAttribute appIDPath,
-    MlirLocation querySite) {
-  FailureOr<DynamicInstanceOp> dynInst = unwrap(index)->getInstance(
-      cast<AppIDPathAttr>(unwrap(appIDPath)), unwrap(querySite));
-  if (failed(dynInst))
-    return MlirOperation{nullptr};
-  return wrap(*dynInst);
-}
-
-MLIR_CAPI_EXPORTED MlirAttribute
-circtMSFTAppIDIndexGetChildAppIDsOf(CirctMSFTAppIDIndex idx, MlirOperation op) {
-  auto mod = cast<hw::HWModuleLike>(unwrap(op));
-  return wrap(unwrap(idx)->getChildAppIDsOf(mod));
-}
-
-MLIR_CAPI_EXPORTED
-MlirAttribute circtMSFTAppIDIndexGetAppIDPath(CirctMSFTAppIDIndex idx,
-                                              MlirOperation fromMod,
-                                              MlirAttribute appid,
-                                              MlirLocation loc) {
-  auto mod = cast<hw::HWModuleLike>(unwrap(fromMod));
-  auto path = cast<msft::AppIDAttr>(unwrap(appid));
-  FailureOr<ArrayAttr> instPath =
-      unwrap(idx)->getAppIDPathAttr(mod, path, unwrap(loc));
-  if (failed(instPath))
-    return MlirAttribute{nullptr};
-  return wrap(*instPath);
 }
