+++ lib/Dialect/ESI/ESIAttributes.cpp
@@ -1,0 +1,27 @@
+//===- ESIAttributes.cpp - Implement ESI dialect attributes -------------===//
+//
+// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#include "circt/Dialect/ESI/ESIAttributes.h"
+
+#include "mlir/IR/Builders.h"
+#include "mlir/IR/BuiltinAttributes.h"
+#include "mlir/IR/DialectImplementation.h"
+#include "llvm/ADT/TypeSwitch.h"
+
+using namespace circt;
+using namespace esi;
+
+#define GET_ATTRDEF_CLASSES
+#include "circt/Dialect/ESI/ESIAttributes.cpp.inc"
+
+void ESIDialect::registerAttributes() {
+  addAttributes<
+#define GET_ATTRDEF_LIST
+#include "circt/Dialect/ESI/ESIAttributes.cpp.inc"
+      >();
+}
