+++ lib/Dialect/MSFT/CMakeLists.txt
@@ -7,7 +7,6 @@
 ##===----------------------------------------------------------------------===//
 
 add_circt_dialect_library(CIRCTMSFT
-  AppID.cpp
   DeviceDB.cpp
   MSFTAttributes.cpp
   MSFTDialect.cpp
